pyproject.toml
[tool.poetry]
name = "songpicker2"
version = "0.1.1"
description = ""
authors = ["Maximilian Wu <me@maxesisn.online>"]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0a9"
build = "^0.1.0"
twine = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
commitizen = "^2.20.4"


[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.1"
tag_format = "$version"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
